+++ cmd/snap-bootstrap/bootstrap/export_test.go
@@ -16,4 +16,4 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
  */
-package recover
+package bootstrap
