+++ cmd/snap-bootstrap/partition/export_test.go
