+++ cmd/snap-bootstrap/partition/partition_test.go
